utils/mainspring/requirements.txt
websockets~=15.0.1
python-dotenv~=1.1.1
requests~=2.32.5
aiohttp~=3.11.0
pyyaml~=6.0.1
steam[client]~=1.4.4
